requirements.txt
pillow
numpy
scikit-learn>=1.4
gradio
matplotlib
matplotlib
